requirements.txt
scipy
torch_geometric
potpourri3d
trimesh
chamferdist
pyyaml
tensorboard
sslap @ git+https://github.com/OllieBoyne/sslap@master

-f https://data.pyg.org/whl/torch-2.0.0+cu118.html
torch_cluster 

-f https://data.dgl.ai/wheels/cu118/repo.html
dgl==1.1.1

